+++ scripts/item.js
@@ -525,6 +525,20 @@ export function setConfig() {
                     }
                 }
             },
+            'Overchannel': {
+                'system': {
+                    'uses': {
+                        'per': null,
+                        'recovery': '',
+                        'value': null
+                    },
+                    'consume': {
+                        'amount': null,
+                        'target': '',
+                        'type': ''
+                    }
+                }
+            }
         },
         'itemConfiguration': {
             'Magic Missile': {